+++ src/main/java/com/example/board/controller/UserController.java
@@ -1,0 +1,41 @@
+package com.example.board.controller;
+
+import com.example.board.dto.request.UserLoginRequestDto;
+import com.example.board.service.UserService;
+import com.example.board.exception.TokenException;
+import com.example.board.util.JwtUtil;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import java.util.HashMap;
+import java.util.Map;
+
+@RestController
+@RequestMapping("/users")
+@RequiredArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+    private final JwtUtil jwtUtil;
+
+    @PostMapping("/login")
+    public ResponseEntity<Map<String, String>> login(@RequestBody UserLoginRequestDto requestDto) {
+        String token = userService.login(requestDto);
+
+        Map<String, String> response = new HashMap<>(); // 토큰을 json 형식으로 전달하기위한 map<>형식 사용
+        response.put("token", token);
+
+        return ResponseEntity.ok(response);
+    }
+
+    @GetMapping("/me")
+    public ResponseEntity<String> getMyInfo(@RequestHeader("Authorization") String authHeader) {
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            throw new TokenException("Authorization 헤더가 잘못됨");
+        }
+
+        String token = authHeader.substring(7);
+        String username = jwtUtil.validateAndGetUserNick(token);
+        return ResponseEntity.ok("안녕하세요, " + username + "님");
+    }
+}
