+++ src/main/java/com/example/board/dto/response/UserResponseDto.java
@@ -1,0 +1,8 @@
+package com.example.board.dto.response;
+
+import lombok.Data;
+
+@Data
+public class UserResponseDto {
+    private String token;
+}
