+++ src/main/java/com/example/board/exception/TokenException.java
@@ -1,0 +1,9 @@
+package com.example.board.exception;
+
+public class TokenException extends RuntimeException {
+	private static final long serialVersionUID = 1L;
+	
+    public TokenException(String message) {
+        super(message);
+    }
+}
