+++ src/main/java/com/example/board/service/impl/CommentServiceImpl.java
@@ -1,0 +1,5 @@
+package com.example.board.service.impl;
+
+public class CommentServiceImpl {
+
+}
