+++ src/main/java/com/example/board/vo/CrudVO.java
@@ -1,0 +1,14 @@
+package com.example.board.vo;
+
+
+import java.time.LocalDateTime;
+
+import lombok.Data;
+
+@Data
+public class CrudVO {
+	private String id;
+	private String title;
+	private String content;
+	private LocalDateTime date;
+}
