--- src/main/java/com/example/jin/MyApplication.java
@@ -1,13 +1,0 @@
-package com.example.jin;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.mybatis.spring.annotation.MapperScan;
-
-@SpringBootApplication
-@MapperScan("com.example.jin.mapper")  // Mapper 인터페이스가 위치한 패키지를 지정
-public class MyApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(MyApplication.class, args);
-    }
-}
