--- src/main/java/com/example/jin/dto/response/BoardResDTO.java
@@ -1,13 +1,0 @@
-package com.example.jin.dto.response;
-
-
-import lombok.Data;
-
-@Data
-public class BoardResDTO {
-	
-	private String boardTitle;				// 게시글 제목
-	private String boardContent;			// 게시글 내용
-	private int boardType;					// 게시글 종류 	ex) 1: 일반 2: 공지
-	private Long userNo;					// 유저 번호
-}
