--- src/main/java/com/example/jin/service/impl/CommentServiceImpl.java
@@ -1,5 +1,0 @@
-package com.example.jin.service.impl;
-
-public class CommentServiceImpl {
-
-}
