--- src/main/java/com/example/joonwoo/dto/UserLoginRequestDto.java
@@ -1,26 +1,0 @@
-package com.example.joonwoo.dto;
-
-import lombok.Data;
-
-@Data
-public class UserLoginRequestDto {
-    private String userId;
-    private String userPw;
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public void setUserId(String username) {
-        this.userId = username;
-    }
-
-    public String getUserPw() {
-        return userPw;
-    }
-
-    public void setUserPw(String password) {
-        this.userPw = password;
-    }
-}
-
