--- src/main/java/com/example/joonwoo/dto/UserResponseDto.java
@@ -1,8 +1,0 @@
-package com.example.joonwoo.dto;
-
-import lombok.Data;
-
-@Data
-public class UserResponseDto {
-    private String token;
-}
