--- src/main/java/com/example/joonwoo/exception/GlobalExceptionHandler.java
@@ -1,39 +1,0 @@
-package com.example.joonwoo.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import java.util.HashMap;
-import java.util.Map;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    // TokenException을 처리
-    @ExceptionHandler(TokenException.class)
-    public ResponseEntity<Map<String, String>> handleTokenException(TokenException ex) {
-        Map<String, String> response = new HashMap<>();
-        response.put("error", ex.getMessage());
-
-        return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
-    }
-
-    // RuntimeException을 처리
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<Map<String, String>> handleRuntimeException(RuntimeException ex) {
-        Map<String, String> response = new HashMap<>();
-        response.put("error", ex.getMessage());
-
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-
-    // 예외 처리
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<Map<String, String>> handleException(Exception ex) {
-        Map<String, String> response = new HashMap<>();
-        response.put("error", "서버 오류가 발생했습니다.");
-
-        return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
