--- src/main/java/com/example/joonwoo/exception/TokenException.java
@@ -1,9 +1,0 @@
-package com.example.joonwoo.exception;
-
-public class TokenException extends RuntimeException {
-	private static final long serialVersionUID = 1L;
-	
-    public TokenException(String message) {
-        super(message);
-    }
-}
