--- src/main/java/com/example/joonwoo/mapper/UserMapper.java
@@ -1,10 +1,0 @@
-package com.example.joonwoo.mapper;
-
-import com.example.joonwoo.entity.UserEntity;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-@Mapper
-public interface UserMapper {
-    UserEntity findByUserId(@Param("userId") String userId);
-}
