+++ src/main/resources/mapper/BoardMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.example.jin.mapper.BoardMapper">
+<mapper namespace="com.example.board.mapper.BoardMapper">
 	
 	<!-- 전체 게시글 조회 -->
 	<select id="getAllList" resultType="BoardVO">
