+++ src/main/resources/mapper/UserMapper.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.example.joonwoo.mapper.UserMapper">
-    <select id="findByUserId" resultType="com.example.joonwoo.entity.UserEntity">
+<mapper namespace="com.example.board.mapper.UserMapper">
+    <select id="findByUserId" resultType="com.example.board.entity.UserEntity">
         SELECT * FROM user WHERE user_id = #{userId} AND is_deleted = false
     </select>
 </mapper>
